Move "type pom struct" from java.go to pom.go.
