+++ cli/azd/internal/appdetect/pom.go
@@ -4,9 +4,111 @@ import (
 	"bufio"
 	"encoding/xml"
 	"fmt"
+	"os"
 	"os/exec"
+	"path/filepath"
+	"regexp"
 	"strings"
 )
+
+// pom represents the top-level structure of a Maven POM file.
+type pom struct {
+	XmlName              xml.Name             `xml:"project"`
+	Parent               parent               `xml:"parent"`
+	Modules              []string             `xml:"modules>module"` // Capture the modules
+	Properties           Properties           `xml:"properties"`
+	Dependencies         []dependency         `xml:"dependencies>dependency"`
+	DependencyManagement dependencyManagement `xml:"dependencyManagement"`
+	Build                build                `xml:"build"`
+	path                 string
+}
+
+// Parent represents the parent POM if this project is a module.
+type parent struct {
+	GroupId    string `xml:"groupId"`
+	ArtifactId string `xml:"artifactId"`
+	Version    string `xml:"version"`
+}
+
+type Properties struct {
+	Entries []Property `xml:",any"` // Capture all elements inside <properties>
+}
+
+type Property struct {
+	XMLName xml.Name
+	Value   string `xml:",chardata"`
+}
+
+// Dependency represents a single Maven dependency.
+type dependency struct {
+	GroupId    string `xml:"groupId"`
+	ArtifactId string `xml:"artifactId"`
+	Version    string `xml:"version"`
+	Scope      string `xml:"scope,omitempty"`
+}
+
+// DependencyManagement includes a list of dependencies that are managed.
+type dependencyManagement struct {
+	Dependencies []dependency `xml:"dependencies>dependency"`
+}
+
+// Build represents the build configuration which can contain plugins.
+type build struct {
+	Plugins []plugin `xml:"plugins>plugin"`
+}
+
+// Plugin represents a build plugin.
+type plugin struct {
+	GroupId    string `xml:"groupId"`
+	ArtifactId string `xml:"artifactId"`
+	Version    string `xml:"version"`
+}
+
+func toPom(filePath string) (*pom, error) {
+	bytes, err := os.ReadFile(filePath)
+	if err != nil {
+		return nil, err
+	}
+
+	var initialProject pom
+	if err := xml.Unmarshal(bytes, &initialProject); err != nil {
+		return nil, fmt.Errorf("parsing xml: %w", err)
+	}
+
+	// replace all placeholders with properties
+	str := replaceAllPlaceholders(initialProject, string(bytes))
+
+	var project pom
+	if err := xml.Unmarshal([]byte(str), &project); err != nil {
+		return nil, fmt.Errorf("parsing xml: %w", err)
+	}
+
+	project.path = filepath.Dir(filePath)
+
+	return &project, nil
+}
+
+func replaceAllPlaceholders(project pom, input string) string {
+	propsMap := parseProperties(project.Properties)
+
+	re := regexp.MustCompile(`\$\{([A-Za-z0-9-_.]+)}`)
+	return re.ReplaceAllStringFunc(input, func(match string) string {
+		// Extract the key inside ${}
+		key := re.FindStringSubmatch(match)[1]
+		if value, exists := propsMap[key]; exists {
+			return value
+		}
+		return match
+	})
+}
+
+func parseProperties(properties Properties) map[string]string {
+	result := make(map[string]string)
+	for _, entry := range properties.Entries {
+		result[entry.XMLName.Local] = entry.Value
+	}
+	return result
+}
 
 func toEffectivePom(pomPath string) (pom, error) {
 	if !commandExistsInPath("java") {
