+++ cli/azd/internal/appdetect/pom_test.go
@@ -1,10 +1,73 @@
 package appdetect
 
 import (
+	"encoding/xml"
 	"os"
 	"path/filepath"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
+
+func TestReplaceAllPlaceholders(t *testing.T) {
+	tests := []struct {
+		name    string
+		project pom
+		input   string
+		output  string
+	}{
+		{
+			"empty.input",
+			pom{
+				Properties: Properties{
+					Entries: []Property{
+						{
+							XMLName: xml.Name{
+								Local: "version.spring-boot_2.x",
+							},
+							Value: "2.x",
+						},
+					},
+				},
+			},
+			"",
+			"",
+		},
+		{
+			"empty.properties",
+			pom{
+				Properties: Properties{
+					Entries: []Property{},
+				},
+			},
+			"org.springframework.boot:spring-boot-dependencies:${version.spring-boot_2.x}",
+			"org.springframework.boot:spring-boot-dependencies:${version.spring-boot_2.x}",
+		},
+		{
+			"dependency.version",
+			pom{
+				Properties: Properties{
+					Entries: []Property{
+						{
+							XMLName: xml.Name{
+								Local: "version.spring-boot_2.x",
+							},
+							Value: "2.x",
+						},
+					},
+				},
+			},
+			"org.springframework.boot:spring-boot-dependencies:${version.spring-boot_2.x}",
+			"org.springframework.boot:spring-boot-dependencies:2.x",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			output := replaceAllPlaceholders(tt.project, tt.input)
+			assert.Equal(t, tt.output, output)
+		})
+	}
+}
 
 func TestToEffectivePom(t *testing.T) {
 	tests := []struct {
